+++ tabs/src/services/lnbitsService.ts
@@ -1,0 +1,1 @@
+/workspace/Zapp.ie/src/components/lnbitsService.ts
